♻️ [devext] extract devtools connection
